requirements.txt
beautifulsoup4==4.12.2
requests==2.29.0
openai==1.3.6
brotlipy==0.7.0
html5lib==1.1
playwright==1.35.0
pandas~=1.1.5